\name{dot-onLoad-lossDev}
\alias{dot-onLoad-lossDev}
\title{Intialize the Namespace.}
\usage{.onLoad(libname, pkgname)}
\description{Intialize the Namespace.  Intended for internal use only.}
\details{Currently only sets correct functions for \code{myPkgName} and \code{myLibPath} and loads the \acronym{JAGS} module.}
\seealso{\code{\link{.onLoad}}}
